Updates with Poplar SDK 3.3.0 release

+++ nlp/gpt2/pytorch/data/indexed_dataset.py
@@ -27,7 +27,7 @@ def logger(msg):
     logging.info(msg)
 
 
-dtypes = {1: np.uint8, 2: np.int8, 3: np.int16, 4: np.int32, 5: np.int64, 6: np.float, 7: np.double, 8: np.uint16}
+dtypes = {1: np.uint8, 2: np.int8, 3: np.int16, 4: np.int32, 5: np.int64, 6: float, 7: np.double, 8: np.uint16}
 
 
 def _warmup_mmap_file(path):
